python-code/2-llmRag/pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = ""
authors = [
    {name = "godinhojoao",email = "godinhojoao2003@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain (==0.3.26)",
    "requests (>=2.32.4,<3.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "tiktoken (>=0.9.0,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

